Removed artificial limitations to RX and TX length by using uint8_t, changed phNxpEse_readPacket to accomodate the 255 byte read limit of wire.h, increased t10i2c_tempBuf size to hold a full frame

+++ SE05X/src/lib/t1oi2c/phNxpEse_Api.h
@@ -37,7 +37,7 @@ typedef enum
  */
 typedef struct phNxpEse_data
 {
-    uint8_t len;     /*!< length of the buffer */
+    uint32_t len;     /*!< length of the buffer */
     uint8_t *p_data; /*!< pointer to a buffer */
 } phNxpEse_data;
